Allow exclusionary tags to omit scenarios
If a user passes in a tag for which no scenario should run, e.g.

   --tag=-A

then the expectation is that no scenario with that tag would run.
However, the previous behavior was to include a scenario if an
inclusion tag matches regardless if an exclusion tag also matches.
scenario with an inclusionary. For example, if a scenario's
tags were:

   ['A', 'B']

and the user specified:

   --tag=-A --tag=B

the scenario would run.

This change changes the behavior of scenario filtering:
1) If there are inclusionary tags and none match, the
   scenario is rejected
2) If there are exclusionary tags and any match, the
   scenario is rejected
3) If there are ~ or ~- tags and none match, the scenario
   is rejected
4) If the above did not reject the scenario, it is accepted

https://github.com/gabrielfalcao/lettuce/issues/498